🗺️ Atlas: [Splitting The Blob: Lexicon and Conversion]
Extracted tests from `src/morphology/lexicon.rs` and `src/semantic/conversion.rs` into sub-modules to reduce file size and improve cohesion.
